+++ trunk/Aeona Sprint 2/Aeona Project/source/Game Objects/CPlayer.cpp
@@ -95,7 +95,8 @@ void CPlayer::Update(float fElapsedTime)
 	
 	// Update the particles
 	m_fxFootsteps.Update(fElapsedTime);
-	LightEngine::GetInstance()->SetPlayerPointPos( ( GetPosX() - ( -1 * CCameraControl::GetInstance()->GetPositionX() ) ), ( GetPosY() - ( -1 * CCameraControl::GetInstance()->GetPositionY() ) ) );
+	LightEngine::GetInstance()->SetPlayerPointPos((float)(GetPosX()-( -1 * CCameraControl::GetInstance()->GetPositionX())),
+		(float)(GetPosY()-( -1 * CCameraControl::GetInstance()->GetPositionY())));
 
 	//WEAPON->SetWeaponRotation(GetAnimationPlayer(GetCurrentAnimation())->ReturnWeaponAngle());
